LayoutTests/imported/w3c/web-platform-tests/css/cssom/getComputedStyle-margins-roundtrip-expected.txt


FAIL Fixed-length margin-left property value should roundtrip assert_equals: expected "20.7px" but got "20.6875px"
FAIL Fixed-length margin-right property value should roundtrip assert_equals: expected "20.7px" but got "20.700001px"
FAIL Fixed-length margin-top property value should roundtrip assert_equals: expected "20.7px" but got "20.6875px"
FAIL Fixed-length margin-bottom property value should roundtrip assert_equals: expected "20.7px" but got "20.6875px"
